Fix moved files
