+++ shared/js/background/classes/tab.js
@@ -13,6 +13,9 @@
  *          }
  *      }
  */
+import { Site } from './site'
+import { HttpsRedirects } from './https-redirects'
+import { TabState } from './tab-state'
 const gradeIconLocations = {
     A: '/img/toolbar-rating-a_48.png',
     'B+': '/img/toolbar-rating-b-plus_48.png',
@@ -25,18 +28,15 @@ const gradeIconLocations = {
     F: '/img/toolbar-rating-f_48.png'
 }
 
-const Site = require('./site.es6')
 const { Tracker } = require('./tracker')
-const HttpsRedirects = require('./https-redirects.es6')
 const Companies = require('../companies.es6')
 const browserWrapper = require('../wrapper.es6')
 const webResourceKeyRegex = /.*\?key=(.*)/
 const { AdClickAttributionPolicy } = require('./ad-click-attribution-policy')
-const { TabState } = require('./tab-state')
 
 /** @typedef {{tabId: number, url: string | undefined, requestId?: string, status: string | null | undefined}} TabData */
 
-class Tab {
+export class Tab {
     /**
      * @param {TabData|TabState} tabData
      */
@@ -327,5 +327,3 @@ class Tab {
         this.ampUrl = url
     }
 }
-
-module.exports = Tab
