+++ shared/js/background/utils.js
@@ -1,7 +1,7 @@
 import browser from 'webextension-polyfill'
 import { getExtensionVersion, getFromSessionStorage, setToSessionStorage } from './wrapper.es6'
 import tdsStorage from './storage/tds.es6'
-import settings from './settings.es6'
+import * as settings from './settings'
 import load from './load.es6'
 import * as tldts from 'tldts'
 import parseUserAgentString from '../shared-utils/parse-user-agent-string.es6'
