+++ unit-test/background/click-to-load-tests.js
@@ -1,5 +1,5 @@
-
 import * as beforeRequest from '../../shared/js/background/before-request'
+import * as settings from '../../shared/js/background/settings'
 require('../helpers/mock-browser-api')
 const chromeWrapper = require('../../shared/js/background/wrapper.es6.js')
 const tds = require('../../shared/js/background/trackers.es6')
@@ -7,7 +7,6 @@ const tdsStorage = require('../../shared/js/background/storage/tds.es6')
 const tdsStorageStub = require('./../helpers/tds.es6')
 const Tab = require('../../shared/js/background/classes/tab.es6')
 const tabManager = require('../../shared/js/background/tab-manager.es6')
-const settings = require('../../shared/js/background/settings.es6')
 
 describe('Tracker Utilities', () => {
     let tabObserver
