+++ unit-test/background/reference-tests/expire-1p-tracking-cookies.js
@@ -1,4 +1,4 @@
-import { handleRequest } from '../../../shared/js/background/before-request'
+import { handleRequest } from '../../../shared/js/background/before-request.es6'
 require('../../helpers/mock-browser-api')
 
 const tds = require('../../../shared/js/background/trackers')
