+++ unit-test/helpers/settings.es6.js
@@ -4,7 +4,7 @@
  * Because this is a Jasmine spy, it resets after every test case,
  * so make sure to have it within a it() or beforeEach()/beforeAll()
  */
-const origSettings = require('../../shared/js/background/settings.es6')
+import * as origSettings from '../../shared/js/background/settings'
 
 const stub = (startingVals) => {
     const settingObj = startingVals || {}
